cleanup depthai script by moving NN handler into different file
