+++ depthai_helpers/age_gender_recognition_handler.py
@@ -1,0 +1,25 @@
+import cv2
+import numpy as np
+
+
+def decode_age_gender_recognition(nnet_packet):
+    detections = []
+    for _, e in enumerate(nnet_packet.entries()):
+        if e[1]["female"] > 0.8 or e[1]["male"] > 0.8:
+            detections.append(e[0]["age"])  
+            if e[1]["female"] > e[1]["male"]:
+                detections.append("female")
+            else:
+                detections.append("male")
+    return detections
+
+def show_age_gender_recognition(entries_prev, frame, **kwargs):
+    # img_h = frame.shape[0]
+    # img_w = frame.shape[1]
+    if len(entries_prev) != 0:
+        age = (int)(entries_prev[0]*100)
+        cv2.putText(frame, "Age: " + str(age), (0, 10), cv2.FONT_HERSHEY_SIMPLEX, 0.5, (0, 0, 255), 2)
+        gender = entries_prev[1]
+        cv2.putText(frame, "G: " + str(gender), (0, 30), cv2.FONT_HERSHEY_SIMPLEX, 0.5, (0, 0, 255), 2)
+    frame = cv2.resize(frame, (300, 300))
+    return frame
